fix(task): change task_hash return type from size_t to int
Fix type conversion warnings in task.c where task_hash() was returning
size_t but being assigned to int idx variables. This caused
-Wshorten-64-to-32 warnings when strict typing was enabled.

Changes:
- task.c: Change task_hash return type from size_t to int
- task.h: Add task_hash function declaration
- signal.c: Update extern declaration from size_t to int

Fixes VAL-EPIC1-001, VAL-EPIC1-002, VAL-EPIC1-003

+++ ixland-system/kernel/signal/signal.c
@@ -223,7 +223,7 @@ int iox_killpg(pid_t pgrp, int sig) {
 
     extern pthread_mutex_t task_table_lock;
     extern iox_task_t *task_table[];
-    extern size_t task_hash(pid_t pid);
+    extern int task_hash(pid_t pid);
 
     pthread_mutex_lock(&task_table_lock);
